Fix deadlock in pointcloud destructor
Internally, the cwipc.pointcloud destructor checks the pointer member is
IntPtr.Zero. However, when the ref count expires, BaseMemoryChunk will
instead throw an exception.

Throwing an exception in a property is a bad idea, and throwing an
exception in a destructor is a disaster. This prevents the situation
from happening.